"Moved BackpackPreview to new sub-package 'item'"
Refactored code by moving the BackpackPreview class to a new 'item' sub-package.
